# openapi.yaml
openapi: 3.0.3
info:
  title: Notify API Platform
  description: |
    Plataforma empresarial API-First para la orquestación, gestión y despacho de notificaciones transaccionales multicanal (Email, SMS y Push).
    Diseñada bajo arquitectura asíncrona resiliente con soporte de JWT, Idempotencia, Rate Limiting y especificación de errores RFC 7807.
  version: 1.0.0
  contact:
    name: Soporte de Integración Notify API
    email: soporte-api@notify.consultoria.com
    url: https://notify.consultoria.com/docs
  license:
    name: Propiedad Privada - Consultoría de Software
    url: https://notify.consultoria.com/terms
servers:
  - url: https://api.notify.consultoria.com/api/v1
    description: Servidor de Producción (AWS us-east-1)
  - url: https://sandbox.notify.consultoria.com/api/v1
    description: Servidor Sandbox / Pruebas (Adaptadores simulados)
...
paths:
...
  /templates:
    get:
      tags:
        - Templates
      summary: Listar plantillas de mensajería registradas
      operationId: listTemplates
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Lista de plantillas disponibles para la aplicación.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TemplateResponse'
    post:
      tags:
        - Templates
      summary: Registrar una nueva plantilla de mensajería
      operationId: createTemplate
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTemplateRequest'
      responses:
        '201':
          description: Plantilla registrada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateResponse'
        '400':
          $ref: '#/components/responses/BadRequestProblem'

  # =========================================================================
  # CHANNELS
  # =========================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Ingrese el token JWT precedido por la palabra Bearer (ej. "Bearer eyJhbGciOi...").
...
  responses:
    BadRequestProblem:
      description: Error de sintaxis o validación de campos en la petición (RFC 7807).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ChannelType:
      type: string
      enum: [EMAIL, SMS, PUSH]
      description: Canal de comunicación soportado.
      example: EMAIL
...
    CreateTemplateRequest:
      type: object
      required:
        - code
        - name
        - channel
        - bodyTemplate
      properties:
        code:
          type: string
          example: WELCOME_USER
        name:
          type: string
          example: Plantilla de Bienvenida
        channel:
          $ref: '#/components/schemas/ChannelType'
        subject:
          type: string
          example: Bienvenido {{nombre}} a Nuestra Plataforma
        bodyTemplate:
          type: string
          example: <h1>Hola {{nombre}}</h1><p>Tu cuenta fue activada.</p>
        requiredVariables:
          type: array
          items:
            type: string
          example: [nombre]
    TemplateResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
          example: WELCOME_USER
        name:
          type: string
          example: Plantilla de Bienvenida
        channel:
          $ref: '#/components/schemas/ChannelType'
        subject:
          type: string
          nullable: true
        version:
          type: integer
          example: 1
        updatedAt:
          type: string
          format: date-time

    # -----------------------------------------------------------------------
    # CHANNEL & AUDIT SCHEMAS
    # -----------------------------------------------------------------------
...
    ProblemDetails:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          description: URI de referencia que categoriza el tipo de error ocurrido.
          example: https://notify.consultoria.com/errors/validation-failed
        title:
          type: string
          description: Resumen corto y legible del tipo de problema.
          example: Error de Validación de Parámetros
        status:
          type: integer
          description: Código de estado HTTP retornado.
          example: 400
        detail:
          type: string
          description: Explicación detallada de la causa específica del fallo.
          example: El destinatario ingresado no corresponde a un formato E.164 válido para SMS.
        instance:
          type: string
          format: uri-reference
          description: Ruta URI donde se originó el fallo.
          example: /api/v1/notifications
        invalidParams:
          type: array
          description: Desglose campo a campo de fallos de validación (exclusivo para 400 Bad Request).
          items:
            $ref: '#/components/schemas/InvalidParamItem'
        traceId:
          type: string
          description: Identificador de correlación para soporte técnico y depuración en logs.
          example: 00-4bf92f3577b34da6a3ce929d0e0e4736-00f067aa0ba902b7-01
    InvalidParamItem:
      type: object
      required:
        - name
        - reason
      properties:
        name:
          type: string
          example: recipient
        reason:
          type: string
          example: El número debe iniciar con prefijo internacional (+) seguido del código de país.